% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_rqs.R
\name{coef_rqs}
\alias{coef_rqs}
\title{Unified Summary for Quantile Regression}
\usage{
coef_rqs(rqs, ...)
}
\arguments{
\item{rqs}{A \code{rq} or \code{rqs} object from \code{{quantreg}}.}

\item{...}{Arguments passed on to \code{\link[quantreg:summary.rq]{quantreg::summary.rq()}}}
}
\value{
A \code{tibble}
}
\description{
A wrapper for \code{summary.rq()}. Display a quantile reregression's call in a tidy format,
one column for each variable and the coefficient, confidence intervals, and tau in the rows.
This makes it easy to make plots with confidence intervals for each coefficient.
}
\seealso{
\code{\link[quantreg:rq]{quantreg::rq()}}
}
