% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortaleza.R
\docType{data}
\name{fortaleza}
\alias{fortaleza}
\title{Fortaleza Neighborhoods}
\format{
\subsection{\code{fortaleza}}{

A simple feature collection with 121 features and 3 fields
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -38.6379 ymin: -3.894603 xmax: -38.40133 ymax: -3.691722}
\item{Geodetic CRS}{SIRGAS 2000}
\item{nome_bairro}{Neighborhood name, withouth capital letters os special characters (character)}
\item{cod_bairro}{Neighborhood code (character)}
\item{area_ha}{Neighborhood area in hectares (numeric)}
\item{geometry}{Neighborhood geometry (sfc_POLYGON)}
}
}
}
\source{
\url{https://mapas.fortaleza.ce.gov.br/}
}
\usage{
fortaleza
}
\description{
Official neighborhoods shapefile of Fortaleza, Brazil.
}
\keyword{datasets}
