% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bbox.R
\name{geom_bbox}
\alias{geom_bbox}
\title{Limit a ggplot map}
\usage{
geom_bbox(data)
}
\arguments{
\item{data}{An object with classes \code{sf} and \code{data.frame}.}
}
\value{
A \code{ggplot} object
}
\description{
This function creates a bounding box from a \code{sf} object and apply its limits to \code{coord_sf}.
}
\examples{
data("fortaleza")

# plot the whole city, fill neighborhood "aldeota" in red, and zoom in to region "SER II"
fortaleza \%>\%
  ggplot() +
  geom_sf(fill = NA) +
  geom_sf(data = . \%>\% filter(name_neigh == "aldeota"), fill = "red") +
  geom_bbox(fortaleza \%>\% filter(name_region == "SER II")) +
  theme_void()

}
