% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_toscale.R
\name{geom_sf_toscale}
\alias{geom_sf_toscale}
\title{Make a geom_sf layer with a fixed scale}
\usage{
geom_sf_toscale(data, dist, col_ref = NULL, ref = NULL, ...)
}
\arguments{
\item{data}{An object with classes \code{sf} and \code{data.frame}. Won't inherit data from previous layer
since to make the scale, it is necessary to calculate a buffer around the centroid.}

\item{dist}{Buffer distance, see \code{sf::st_buffer}}

\item{col_ref}{\verb{<data-masking>} Column in \code{data} with the filtering reference.
Optional, provide only if desired output is a subset of original data.}

\item{ref}{Filtering reference to be passed on \code{col_ref}. Either a single value or a vector.}

\item{...}{Other arguments to pass to \code{ggplot2::geom_sf()}.}
}
\value{
A \code{ggplot} object
}
\description{
A wrapper around \code{ggplot2::geom_sf}, this function generates fixed scale maps, which enhances
comparison between different plots.
}
\examples{
data("fortaleza")

# entire data
ggplot() +
  geom_sf_toscale(fortaleza, 15000)

# filtering a few neighborhoods, all at once
ggplot() +
  geom_sf_toscale(fortaleza, 3000, name_neigh, c("centro", "aldeota"))

# filtering a few neighborhoods, one per plot
p <- fortaleza[1:2,] \%>\%
  pull(name_neigh) \%>\%
  map(
    \(x) ggplot() +
      geom_sf(data = fortaleza, fill = "white") +
      geom_sf_toscale(fortaleza, 2000, name_neigh, x) +
      labs(title = x)
  )

## plotting them together
cowplot::plot_grid(p[[1]], p[[2]])
}
