% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm.R
\name{get_osm_roads}
\alias{get_osm_roads}
\alias{get_osm_zips}
\alias{get_osm_rail}
\title{Download osm data}
\usage{
get_osm_roads(bbox)

get_osm_zips(bbox)

get_osm_rail(bbox)
}
\arguments{
\item{bbox}{A bounding box for the search area.}
}
\value{
A \code{list} for each kind of geometry retrieved by osmdata.
}
\description{
A convenient wrapper for osm features to get some features I need easily convertible to \code{sf} format.
}
\details{
Returns values that meet \code{key} = "highway" and \code{value} = "motorway", "primary", "secondary",
"tertiary", "residential", "living_street" and "pedestrian". For more info, please check
\href{https://wiki.openstreetmap.org/wiki/Key:highway}{"Key:highway"} page at OSM Wiki.

Returns values that meet \code{key} = "addr:postcode". For more info, please check
\href{https://wiki.openstreetmap.org/wiki/Key:railway}{"Key:railway"} page at OSM Wiki.

Returns values that meet \code{key} = "railway" and \code{value} = "rail", "light_rail", "subway", "tram",
"service", "construction", "usage", "station", and "tram_stop". For more info, please check
\href{https://wiki.openstreetmap.org/wiki/Key:addr:*}{"Key:addr:*"} page at OSM Wiki
}
\examples{
\dontrun{
# Get roads in Brazil's smallest city
geobr::read_municipality(code_muni = 3157336) \%>\%
sf::st_bbox() \%>\%
get_osm_roads()
}
\dontrun{
# Get postcodes in Brazil's smallest city
geobr::read_municipality(code_muni = 3157336) \%>\%
sf::st_bbox() \%>\%
get_osm_zips()
}
\dontrun{
# Get railways in Brazil's smallest city
geobr::read_municipality(code_muni = 3157336) \%>\%
sf::st_bbox() \%>\%
get_osm_rail()
}
}
\seealso{
\code{\link[osmdata:add_osm_feature]{osmdata::add_osm_feature()}}
}
