% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm.R
\name{get_osm_rail}
\alias{get_osm_rail}
\title{Download osm rail transit}
\usage{
get_osm_rail(bbox)
}
\arguments{
\item{bbox}{A bounding box for the search area.}
}
\value{
Object with classes \code{list}, \code{osmdata}, and \code{osmdata_sf}
for each kind of geometry retrieved by osmdata.
}
\description{
A convenient wrapper for osm features to get (rail) transit lines and stations
easily convertible to \code{sf} format
}
\details{
Returns values that meet \code{key} = "railway" and \code{value} = "rail", "light_rail", "subway", "tram",
"service", "construction", "usage", "station", and "tram_stop". For more info, please check
\href{https://wiki.openstreetmap.org/wiki/Key:addr:*}{"Key:addr:*"} page at OSM Wiki
}
