% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm.R
\name{get_osm_roads}
\alias{get_osm_roads}
\title{Download osm roads}
\usage{
get_osm_roads(bbox)
}
\arguments{
\item{bbox}{A bounding box for the search area.}
}
\value{
A \code{list} for each kind of geometry retrieved by osmdata.
}
\description{
A convenient wrapper for osm features to get roads easily convertible to \code{sf} format.
}
\details{
Returns values that meet \code{key} = "highway" and \code{value} = "motorway", "primary", "secondary",
"tertiary", "residential", "living_street" and "pedestrian". For more info, please check
\href{https://wiki.openstreetmap.org/wiki/Key:highway}{"Key:highway"} page at OSM Wiki.
}
