% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm.R
\name{get_osm_zips}
\alias{get_osm_zips}
\title{Download osm postcodes}
\usage{
get_osm_zips(bbox)
}
\arguments{
\item{bbox}{A bounding box for the search area.}
}
\value{
A \code{list} for each kind of geometry retrieved by osmdata.
}
\description{
A convenient wrapper for osm features to get postcodes easily convertible to \code{sf} format
}
\details{
Returns values that meet \code{key} = "addr:postcode". For more info, please check
\href{https://wiki.openstreetmap.org/wiki/Key:railway}{"Key:railway"} page at OSM Wiki.
}
