% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rais.R
\name{get_rais}
\alias{get_rais}
\title{Download RAIS data}
\usage{
get_rais(variables, source, city, geo_filter, ano, path)
}
\arguments{
\item{variables}{A character vector containing the selected variables. To select all, use
variables = "*".}

\item{source}{Data source in BigQuery.}

\item{city}{IBGE 7-digit code of the city in single quotes, e.g. "id_municipio = '1234567'.}

\item{geo_filter}{Optional geographic filters for districts and neighborhoods, when available.}

\item{ano}{Selected year(s).}

\item{path}{Path, including file name and extension, to save results to.}
}
\value{
a \code{.csv} file downloaded in the specified path
}
\description{
Get data from the official Brazilian employer-employee dataset (RAIS) using package \code{{basedosdados}}
}
\details{
A simple wrapper of the \code{download()} function to automate download for multiple years. \code{variables},
\code{source}, \code{city}, \code{geo_filter}, and \code{ano} should all be written as part of an SQL query. To select
multiple cities at once, use format \code{city} = "id_municipio IN (city1, city2)". For more information
on the RAIS dataset available at \code{{basedosdados}}, please refer to
https://basedosdados.org/dataset/3e7c4d58-96ba-448e-b053-d385a829ef00?table=c3a5121e-f00d-41ff-b46f-bd26be8d4af3.
}
\examples{
\dontrun{
# Getting RAIS data for Caete (MG)

## setting variables
variables <- "ano, quantidade_vinculos_ativos, cnae_2, indicador_atividade_ano"
source <- "`basedosdados.br_me_rais.microdados_estabelecimentos`"
city <- "id_municipio='3110004'"

## download
get_rais(
  variables = variables,
  source = source,
  city = city,
  ano = 2018,
  path = "data/csv/df_rais_caete.csv"
  )
}
}
