% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphub_to_sf.R
\name{maphub_tidy_stations}
\alias{maphub_tidy_stations}
\title{Tidy transit stations from Maphub.}
\usage{
maphub_tidy_stations(data)
}
\arguments{
\item{data}{A \code{sf} object imported from MapHub and already converted using maphub_to_sf}
}
\value{
A \code{sf} object.
}
\description{
Tidy transit stations from Maphub.
}
\details{
Very context-specific, columns created using maphub_to_sf() should match colum names here.
}
