% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphub_to_sf.R
\name{maphub_to_sf}
\alias{maphub_to_sf}
\title{Widen description from MapHub}
\usage{
maphub_to_sf(data, values = description, key_sep = ": ", pair_sep = "\\n")
}
\arguments{
\item{data}{A \code{sf} object imported from MapHub and read into R using \code{sf::st_read}.}

\item{values}{\if{html}{\out{<data-masking>}} Column containing data to be converted into multiple variables.
Defaults to \code{description}.}

\item{key_sep}{Character string that separates a key from its value.}

\item{pair_sep}{Character string that separates key-value pairs.}
}
\value{
A \code{sf} object.
}
\description{
Convert maphub geojson-like description info to tibble columns.
}
\details{
Returns a spatial dataset with separate columns for each key-value pair in the original description
column
}
