% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_extract_read.R
\name{shp_extract_read}
\alias{shp_extract_read}
\title{Extract and read compacted shapefiles}
\usage{
shp_extract_read(path, dsn, ...)
}
\arguments{
\item{path}{Path of the compacted folder, including file name and extension.}

\item{dsn}{data source name (interpretation varies by driver - for some
drivers, \code{dsn} is a file name, but may also be a folder, or contain
the name and access credentials of a database); in case of GeoJSON,
\code{dsn} may be the character string holding the geojson data. It can
also be an open database connection.}

\item{...}{
  Arguments passed on to \code{\link[sf:st_read]{sf::st_read}}
  \describe{
    \item{\code{layer}}{layer name (varies by driver, may be a file name without
extension); in case \code{layer} is missing, \code{st_read} will read the
first layer of \code{dsn}, give a warning and (unless \code{quiet = TRUE})
print a message when there are multiple layers, or give an error if there
are no layers in \code{dsn}. If \code{dsn} is a database connection, then
\code{layer} can be a table name or a database identifier (see
\code{\link[DBI]{Id}}). It is also possible to omit \code{layer} and rather
use the \code{query} argument.}
  }}
}
\value{
A \code{sf} object
}
\description{
Extract shapefiles from a compacted folder and read them using \code{sf::st_read}.
}
