% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialops-package.R
\docType{package}
\name{spatialops-package}
\alias{spatialops}
\alias{spatialops-package}
\title{spatialops: Spatial operations that comes in handy}
\description{
Mainly spatial operations, like calculating distances or classifying catchment areas, but also handy tools like ggplot wrappers.
}
\author{
\strong{Maintainer}: Arthur Bazolli \email{baarthur0@outlook.com}

}
\keyword{internal}
