% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_abnq.R
\name{theme_abnq}
\alias{theme_abnq}
\title{ggplot in ABNT format}
\usage{
theme_abnq(
  base_size = 10,
  base_family = "Times",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Font baseline size. Defaults to 10pt.}

\item{base_family}{Font family. Defaults to "Times" (for Times New Roman).}

\item{base_line_size, base_rect_size}{Base stroke for lines and rectangles. Defaults to dividing
the font baseline size by 22 (\code{ggplot}'s default).}
}
\value{
A \code{ggplot} object
}
\description{
\code{{abnquarto}} template for \code{ggplot} objects.
}
\examples{
\dontrun{
data(mtcars)
p <- mtcars \%>\%
  ggplot() +
  geom_point(
    aes(x = hp, y = mpg, color = factor(am))
    ) +
 labs(
    title = "The fast and the goulash",
    subtitle = "The statistical analysis!",
    caption = "Ye old dummy",
    tag = "Tag along",
    x = "Horsepower",
    y = "Miles per gallon"
    ) +
 scale_color_discrete(name = "I'm a legend") +
 facet_wrap(vars(gear))

 p + theme_abnq()
}
}
