% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_abnq.R
\name{theme_abnq_map}
\alias{theme_abnq_map}
\title{ggplot map in ABNT format}
\usage{
theme_abnq_map(base_size = 10, base_family = "Times")
}
\arguments{
\item{base_size}{Font baseline size. Defaults to 10pt.}

\item{base_family}{Font family. Defaults to "Times" (for Times New Roman).}
}
\description{
\code{{abnquarto}} template for \code{ggplot}. Void version to use with maps.
}
\examples{
\dontrun{
data(mtcars)
p <- mtcars \%>\%
  ggplot() +
  geom_point(
    aes(x = hp, y = mpg, color = factor(am))
    ) +
 labs(
    title = "The fast and the goulash",
    subtitle = "The statistical analysis!",
    caption = "Ye old dummy",
    tag = "Tag along",
    x = "Horsepower",
    y = "Miles per gallon"
    ) +
 scale_color_discrete(name = "I'm a legend") +
 facet_wrap(vars(gear))

 p + theme_abnq_map()
}
}
