% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinary_sets.R
\name{trinary_sets}
\alias{trinary_sets}
\title{Trinary sets}
\usage{
trinary_sets(df, cat1, cat2, keep.overlap = F)
}
\arguments{
\item{df}{A \code{sf} object.}

\item{cat1}{\if{html}{\out{<data-masking>}} Column containing the object classes. Cannot contain more than three
distinct classes.}

\item{cat2}{\if{html}{\out{<data-masking>}} Second category to guide sets classification.}

\item{keep.overlap}{Should overlapping sets (i.e. entire sets and unions) be included? Defaults
to \code{FALSE}.}
}
\value{
A \code{sf} object
}
\description{
Perform spatial operations in sets of three classes and an additional category.
}
\details{
Returned data contains one geometry for each class (A', B', C', AB, AC, BC, ABC, and overlaps
when required) per additional category. Categories are renamed according to the order displayed
in the message after the code runs successfully.
}
