% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_rais.R
\name{geom_sf_rais}
\alias{geom_sf_rais}
\title{RAIS thematic maps using ggplot}
\usage{
geom_sf_rais(
  x,
  scale_theme = c("distiller", "viridis"),
  palette = NULL,
  direction = 1,
  data_format = c("absolute", "relative"),
  data_level = c("level", "growth"),
  auto.labels = TRUE,
  geo_name = NULL,
  geo_level = NULL,
  var_name = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  faceted = FALSE,
  facet_var = NULL,
  faceted_labs = NULL
)
}
\arguments{
\item{x}{A \code{data.frame} to plot or an object convertible to it.}

\item{scale_theme}{Palette theme to be used in the plots, currently "distiller" and "viridis" are supported}

\item{palette}{Optional palette specification. For absolute values, defaults to "YlGnBu" when
\code{scale_theme} = "distiller" or "D" if \code{scale_theme} = "viridis". For relative values, defaults to "RdBu".}

\item{direction}{Palette's direction, defaults to 1.}

\item{data_format}{Specify if data is in absolute (default) or relative format.}

\item{data_level}{Specify if data represents a point value or growth between two periods.}

\item{auto.labels}{When \code{TRUE}, automatically generates title, subtitle, and caption based on \code{geo_name},
\code{geo_level}, and \code{var_name}. If \code{FALSE}, specify \code{title}, \code{subtitle}, and \code{caption}.}

\item{geo_name}{Name of the geographic area being represented.}

\item{geo_level}{Name of the smallest geographic area in the data (e.g. state, neighborhood).}

\item{var_name}{Name of the plotted variable.}

\item{title}{Plot title.}

\item{subtitle}{Plot subtitle.}

\item{caption}{Plot caption.}

\item{faceted}{Defaults to \code{FALSE}, change to \code{TRUE} to use a faceted plot.}

\item{facet_var}{A variable to be passed into \code{ggplot2::vars()} when faceting.}

\item{faceted_labs}{Labels for the faceted subplots.}
}
\value{
A \code{ggplot} object
}
\description{
RAIS thematic maps using ggplot
}
